Add option to wait for Schema to be ready.
Current behavior waits for the entire input PCollection to be read
and closed before reading the Schema. This can delay the pipeline
for large inputs, and does not guarantee that the schema is ready
(if it is created in the same pipeline) for small inputs.

This adds an optional parameter with a PCollection to wait for
before reading the schema. If not specified, the schema is read
immediately.