[BEAM-3089] Use Flink cluster parallelism if no parallelism provided
The Runner always defaulted to 1, even if the Flink cluster had a default
parallelism set. With this patch, when no parallelism has been provided, the
cluster default will be used.

This was working before, it is a regression of
cdd2544ba6dd6ac4aa80c65ecd8e01ab3cf664aa.