Gather all mutations in a bundle, sort and batch
In batch mode bundles can be large, in streaming mode
it depends on the load, so this effectively adapts the batching
depending on the load and mode.

MutationGroupEncoder class is renamed to MutationKeyEncoder, as it no
longer encodes entire MutationGroups, only the Key. (SerializableCoder
is now used to encode entire MutationGroups).

SerializedMutation is not longer required as serialized Mutation Groups
are passed between transforms as byte[]

Also adds diagnostic counters.