[BEAM-5442] Add basic FlinkOptions to Python SDK
This adds two common FlinkOptions to the Python SDK which are otherwise filtered
out when using the PortableRunner. It makes sense to validate known options and
we should eventually add all options in FlinkPipelineOptions.

In a follow-up we should ensure that unknown parameters are also forwarded to
the Runner which can then decide what to do with them.