[BEAM-5442] Pass unknown SDK pipeline options to Runner
Unknown Python SDK pipelin options were previously ignored. Pipeline options had
to be replicated in the Python SDK to be passed on to the Runner.

Validating known options in the SDK is a good way to report failures back to the
user early. However, it also requires to maintain options in multiple places.

This change passes on unknown options as long as they are in the form of
'--option value' or '--option=value'. It logs whenever that happens. All other
options are discarded, as previously the case. Discarding option is now logged
with a warn. For example:

```
INFO:root:Parsing unknown args: ['-invalid=value', '--valid=value']
WARNING:root:Discarding unparseable args: ['-invalid=value']
```