[BEAM-5366] Vendor gRPC and its transitive dependencies as a separate package
This prevents beam-model-pipeline, beam-model-fn-execution and beam-model-job...
from having copies of gRPC and their transitive dependencies being published
as artifacts within Apache Beam.

Note that this is known to break the Dataflow import process since we
are defining a new Gradle project and can only be fixed up during
the actual import.