Don't redundantly execute pre-commits as part of Jenkins post-commits
Each post-commit suite is configured to also include pre-commit tests.
However, in Jenkins we run pre-commits separately from post-commits, and
in all cases when a language post-commit suite runs, the pre-commit is
also run separate (in PR, on merge, cron schedule). So, it makes sense
to separate the targets. This will free up resources and give a faster
signal on post-commit suites as they are doing less work.