Remove Gradle from Jenkins job names
This is an artifact from the build migration from Maven to Gradle.
During migration we were maintaining both build systems, and had Jenkins
jobs for both variants. Now that Maven is gone, there is no need to
specify 'Gradle' in the job name. This makes all the job names a bit
shorter.