[BEAM-5386] Prevent CheckpointMarks from not getting acknowled
The UnboundedSourceWrapper would still shutdown when the final watermarks of all
readers are equal to the maximum timestamp. This could then lead to the remaining
CheckpointMarks to never get acknowledged. Acknowledgement is performed as part
of the checkpoint completion which is only possible when the source and its task
are running.

A user had previously reported that the last remaining messages were not getting
acknowledged in a PubSub scenario. It looks like this is the cause.