[BEAM-4681] Address synchronization issue for portable timers
This fixes a regression of BEAM-867 where the key needs to be set on the state
backend to support deleting pending timers on timer registration or firing. This
can intefere with accessing user-defined state, but could also interfere when a
timer is set at the same time that a timer fires.