Replace visteg plugin with gradle-task-tree to visualize task dependency tree
Visualizing the Gradle task dependency tree is a useful tool for
debugging build issues. We previously used the visteg plugin, although
the project seems to be abandoned and not compatible with Gradle 5.0

This change switches over to using the gradle-task-tree plugin, which
gives similar functionality (although only prints the tree to the
commandline rather than a .dot file).

Usage:

  ./gradlew :myProject:myTask :myProject:taskTree

See: https://github.com/dorongold/gradle-task-tree