[BEAM-6510] Use TestFlinkRunner for Nexmark tests
In streaming mode, the FlinkRunner does not shutdown after executing a pipeline
because checkpointing does not work correctly otherwise. There was logic in the
past that did an auto-shutdown in case checkpointing was disabled. The logic has
been removed because it caused confusion for users.

The `TestFlinkRunner` still has the auto-shutdown logic in place.