[BEAM-5638] Exception handling for Java MapElements and FlatMapElements
Previously attempted in #6586, this incorporates feedback from the previous
review and provides some additional convenience. The most important change
is that the interface now expects users to provide an exception handler
function that outputs elements that are efficiently codable as originally
suggested by @tims.

Common features have been implemented in a new `WithExceptions` class,
and the top-level documentation there is the best place to look for an overview
of the approach here.

This PR does not address Filter, AsJsons, and ParseJsons. Those could be added
in a future PR.