[BEAM-6650] Convert FlinkKeyGroupStateInternals to using KeyedStateFunction
The previous implementation was not used anymore which caused elements emitted
during finalization of the bundle to be excluded from the checkpoint. This
resulted in lost elements.

The new implementation uses keyed state which is automatically checkpointed. The
KeyedStateFunction is applied on the state backend to extract and flush all the
elements after a restore.