[BEAM-6679] Fix javadoc @ literals
This change ensures that the core module javadoc:

    ./gradlew ':beam-sdks-java-core:javadoc'

no longer produces the string `@literal` anywhere in the produced html.
There may still be some instances in other modules.

Many code examples in the core SDK use `{@literal @}` inside of a `{@code}`
block, which does escape the `@` as expected, but rather produces the text
`{@literal @}`.

There's unfortunately not any single solution that's elegant in all cases.
The SDK already uses a variety of markup styles for code examples, and I expect
that to continue to be the case depending on the complexity of the example.

I've made additional whitespace fixes along with changing how `@literal` is used
such that each of these code examples now renders nicely with the current
Java 8 build.