[BEAM-6810] Disable CalcRemoveRule
Trivial programs project precisely their input fields, without dropping
or re-ordering them. It turns out that CalcRemoveRule eliminates them,
even if there is a field aliasing. It results in an unexpected
resulting schema of SQL query.

It isn't clear whether it's an issue in Calcite or Beam rules, as a
workaround we disable CalcRemoveRule.