[BEAM-6770] Correct zstd-jni dependency scope to tests only
As discussed in
[the Jira ticket](https://issues.apache.org/jira/browse/BEAM-6770),
we are moving this dependency to be test-only by default and documenting
that it's the user's responsibility to explicitly declare a dependency on
`zstd-jni` before attempting to read or write .zst files.

`zstd-jni` is effectively now an optional dependency, even though it won't
be explicity listed as such in the produced pom. The cost:benefit ratio
of codifying an optional dependency in the build is considered low
at this point.