[BEAM-3863] Ensure correct firing of processing time timers
In Beam, a timer with timestamp `T` is only illegible for firing when the time
has moved past this time stamp, i.e. `T < current_time`. In the case of event
time, current_time is the watermark, in the case of processing time it is the
system time.

Flink's TimerService has different semantics because it only ensures `T <=
current_time`. To make up for this, we previously subtracted one from the
Watermark. However, this does not fix the problem for processing time. We can't
modify processing time easily like we can for event time via the watermark.

To fix processing timers, we change Flink's internal timer timestamp. We add one
millisecond to the timestamp to ensure that the time has moved past the original
timer timestamp in both event and processing time.

Note that we do not modify Beam's timestamp and we do not expose Flink's
timestamp. For consistency this approach has also been applied to event time
timers.