[BEAM-7126] Remove duplicate key encoding for state backend
The key for state requests in portable pipelines already comes encoded as a byte
array. We just have to ensure that the same encoding is used when setting
timers for the cleanup to work correctly. There are tests for this already and
this adds another check if the state has been removed from the state backend.