[BEAM-7366] Determine Spotless target based on project's SourceSet
Spotless currently traverses the project files to discover the source
files. This can causes problems when the project directory uses a source set
which spans outside the project folder. This is the case for the Flink Runner
projects which pull in their main source from `runners/flink` with the version
specific code being in `runners/flink/1.x`.