[BEAM-8011] Remove workaround to run unbounded sources in portable pipelines
In the course of BEAM-6730 we added a way to Runner-translate Reads in portable
pipelines. GreedyPipelineFuser was manipulated to let through Reads to the
Runner. The better solution for this problem would have been to remove the
environment from unbounded reads, which will make it a Runner translated
transform.

If in the future we support unbounded reads via a SDF adapter, we can re-add the
environment.