[BEAM-7760] Added Interactive Beam module
1. Added interactive_beam module that will serve sugar syntax and
shorthand functions to apply interactivity and visualize
PCollection data.
2. This commit implemented the implicitly managed Interactive Beam
environment to track definition of user pipelines. It exposed a watch()
interface for users to explicitly instruct Interactive Beam the
whereabout of their pipeline definition when it's not in __main__.
3. This commit exposed a PCollection data exploration interface
visualize(). Implementation is yet to be added.
4. Added interactive_environment module for internal usage without
backward-compatibility. It holds the cache manager and watchable
metadata for current interactive environment/session/context. Interfaces
are provided to interact with the environment and its components.
5. Unit tests included.