[BEAM-7760] Added Interactive Beam module
1. Added interactive_beam module that will serve sugar syntax and
shorthand functions to apply interactivity, create pipeline with
interactivity, visualize PCollection data and execute pipeline
with interactivity as normal non-interactive pipeline with
selected Beam runners and pipeline options.
2. This commit implemented the implicitly managed Interactive Beam
environment to track definition of user pipelines. It exposed a watch()
interface for users to explicitly instruct Interactive Beam the
whereabout of their pipeline definition when it's not in __main__.
3. This commit implemented a shorthand function create_pipeline()  to
create a pipeline that is backed by direct runner with interactivity
when running.
4. This commit also implemented a shorthand function run_pipeline() to
run a pipeline created with interactivity on a different runner and
pipeline options without interactivity. It's useful when interactivity
is not needed and a one-shot in production-like environment is desired.
5. This commit exposed a PCollection data exploration interface
visualize(). Implementation is yet to be added.
6. Added interactive_environment module for internal usage without
backward-compatibility. It holds the cache manager and watchable
metadata for current interactive environment/session/context. Interfaces
are provided to interact with the environment and its components.
7. Unit tests included.