[BEAM-8157] Remove length prefix from state key for Flink's state backend
The Flink runner requires the internal key to be encoded without a length
prefix (OUTER context). The user state request handler exposes a serialized
version of the key to the Runner. This key is encoded with the NESTED
context. We need to convert it to OUTER context to match the Flink runner's key
encoding.

So far this has not caused the Flink Runner to behave incorrectly. However, with
the upcoming support for Flink 1.9, the state backend will not accept requests
for keys not part of any key group/partition of the operator. This is very
likely to happen with the encoding not being consistent.