[GoSDK] Improve singleton side input runtime error
The Runtime Error message left a bit to be desired. The main thing that's missing for these runtime errors is how to localize this better WRT where in the pipeline graph it's being used (eg. Scope hierarchy or similar), but many DoFns are 1 off, so this might be sufficient.

Before:
making side input 0:
        singleton side input Singleton for *flumego_test_go_proto.KV ill-defined

making side input 0 for fully/qualified/path/to/dofn/create.sideKV:
        got 2 values, want one value for Singleton side input of type *flumego_test_go_proto.KV